Move hotkeys into +layout and refactor into object
